pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "laser_gates"
version = "0.2.0"
description = "A port of the Imagic game using Arcade 3.x and ArcadeActions"
authors = [
    {name = "Brandon Corfman", email = "teacup_canny.5l@icloud.com"}
]
license = {file="LICENSE"}
readme = "README.md"
requires-python = ">= 3.13"
dependencies = [
    "arcade-actions>=0.7",
]

[tool.uv]
dev-dependencies = [
    "arcade-actions",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
]

[tool.uv.sources]
arcade-actions = { path = "../arcade_actions", editable = true }

[tool.pytest.ini_options]
addopts = "--cov-branch --cov-report term --cov=laser_gates"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
markers=["unit: Unit tests.",
         "system: System (browser-based) tests."]
         
[tool.coverage.run]
source = ["laser_gates"]
omit = [
    "src/laser_gates/app.py",
    "src/laser_gates/contexts.py",
    "src/laser_gates/player.py",
    "src/laser_gates/view.py",
    "src/laser_gates/waves.py",
]

[tool.coverage.report]
omit = [
    "src/laser_gates/app.py",
    "src/laser_gates/contexts.py",
    "src/laser_gates/player.py",
    "src/laser_gates/view.py",
    "src/laser_gates/waves.py",
]

[tool.ruff]
line-length = 120
fix = true
target-version = "py313"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]
extend-fixable = ["E", "F", "UP", "B", "SIM", "I"]
